Simplify calculatorState
hasDecimal can be computed from value on demand. Hence, operand
can store the value directly. This simplified several parts of
the code.

This change also fixes a bug: if a decimal point was entered and
then backspaced, it could not be entered again without clearing
the calculator. (The hasDecimal attribute was not properly
updated upon change -- the reason for this commit.)